examples/systems/travel-agent/requirements.txt
openai>=1.0
requests>=2.31
tavily-python>=0.5
python-dotenv>=1.0.0
